requirements.txt
anthropic
snowflake-connector-python
python-dotenv
tabulate